Compress the URL for linking to specific boards
Previously every cell was represented in the string, so for a 30x50
board there were 1500 digits. 0 was dead, 1 was alive.

The new format alternates between white and black with every digit and the
digit itself represents how many of the color were in a run.

So 000011100101 in the old format becomes 0432111. Note the leading 0,
this is because the first digit is the run of living cells.

To save more space, rather than using just numbers, we are using base 36,
for all alphanumeric characters. So 50 zeroes followed by 17 ones in the
old format compresses to z0fh.

As far as I can tell this should make the string ALWAYS shorter, with two
exceptions. An alternating checkerboard starting with black is the same
number of digits long, but looks like 111111... instead of 1010101010...
However, if the checkerboard starts with white, it looks like 0111111...
so we actually gain one digit... oh well :)